Refactoring - Changed KeyboardPressHandler.java to SnakeDirectionInputHandler.java - Changed HelloApplication.java to SnakeGameApplicationInitializer.java - Removed magic numbers in SnakeGameApplicationInitializer and created final variables holding values

+++ src/main/java/com/sdm/snake/Snake.java
@@ -21,7 +21,7 @@ public class Snake {
      * @param length    the initial length of the snake; must be greater than 0.
      * @throws IllegalArgumentException if the length is less than or equal to 0.
      */
-    public Snake(int startingX, int startingY, int length) {
+    public Snake(double startingX, double startingY, int length) {
         this.body = new SnakeBody(startingX, startingY, length);
         this.direction = new SnakeDirection(Direction.UP);
         this.mover = new SnakeMover(this);
